ci: setup minio env
